feat: improved transaction components
