--- src/components/transaction/messages/profiles/MessageSaveProfileListItem.tsx
@@ -1,30 +1,0 @@
-import { MsgSaveProfileEncodeObject } from '@desmoslabs/desmjs';
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import { View } from 'react-native';
-import { Typography } from '../../../typography';
-import { BaseMessageListItem } from '../BaseMessageListItem';
-
-export type Props = {
-  message: MsgSaveProfileEncodeObject["value"];
-  date: Date;
-};
-
-/**
- * Displays the short details of a MsgSaveProfile within a list.
- * @constructor
- */
-export const MessageSaveProfileListItem: React.FC<Props> = ({message, date}) => {
-  const { t } = useTranslation();
-  return (
-    <BaseMessageListItem
-      icon={require('../../../../assets/tx-icons/general.png')}
-      date={date}
-      renderContent={() => (
-        <View>
-          <Typography.Body1>{t('tx type save profile')}</Typography.Body1>
-        </View>
-      )}
-    />
-  );
-};
